CP-54224 Move lvutil.py from /opt
