+++ Makefile
@@ -34,6 +34,7 @@ SM_CORE_LIBS += libiscsi
 
 SM_LIBS := cleanup
 SM_LIBS += mpathcount
+SM_LIBS += lvutil
 
 # Things used as commands which install in libexec
 # which are in python and need compatibility symlinks from
